Refactor self profile downloading
