+++ site/src/self_profile.rs
@@ -1,8 +1,13 @@
 //! This module handles self-profile "rich" APIs (e.g., chrome profiler JSON)
 //! generation from the raw artifacts on demand.
 
+use crate::api::self_profile::ArtifactSize;
+use crate::api::{self_profile, ServerResult};
+use crate::load::SiteCtxt;
 use anyhow::Context;
 use bytes::Buf;
+use database::ArtifactId;
+use std::time::Duration;
 use std::{collections::HashMap, io::Read, time::Instant};
 
 mod codegen_schedule;
@@ -115,6 +120,116 @@ pub(crate) async fn get_self_profile_raw_data(url: &str) -> anyhow::Result<Vec<u
     );
 
     extract(&compressed)
+}
+
+#[derive(Clone)]
+pub struct SelfProfileWithAnalysis {
+    pub profile: self_profile::SelfProfile,
+    pub profiling_data: analyzeme::AnalysisResults,
+}
+
+pub(crate) async fn download_and_analyze_self_profile(
+    ctxt: &SiteCtxt,
+    aid: ArtifactId,
+    benchmark: &str,
+    profile: &str,
+    scenario: database::Scenario,
+    metric: Option<f64>,
+) -> ServerResult<SelfProfileWithAnalysis> {
+    let conn = ctxt.conn().await;
+    let aids_and_cids = conn
+        .list_self_profile(aid.clone(), benchmark, profile, &scenario.to_string())
+        .await;
+
+    let Some((anum, cid)) = aids_and_cids.first() else {
+        return Err(format!("no self-profile found for {aid}"));
+    };
+    let profiling_data =
+        match fetch_raw_self_profile_data(*anum, benchmark, profile, scenario, *cid).await {
+            Ok(d) => extract_profiling_data(d)
+                .map_err(|e| format!("error extracting self profiling data: {}", e))?,
+            Err(e) => return Err(format!("could not fetch raw profile data: {e:?}")),
+        };
+    let profiling_data = profiling_data.perform_analysis();
+    let profile =
+        get_self_profile_data(metric, &profiling_data).map_err(|e| format!("{}: {}", aid, e))?;
+    Ok(SelfProfileWithAnalysis {
+        profile,
+        profiling_data,
+    })
+}
+
+fn get_self_profile_data(
+    cpu_clock: Option<f64>,
+    profile: &analyzeme::AnalysisResults,
+) -> ServerResult<self_profile::SelfProfile> {
+    let total_time: Duration = profile.query_data.iter().map(|qd| qd.self_time).sum();
+
+    let query_data = profile
+        .query_data
+        .iter()
+        .map(|qd| self_profile::QueryData {
+            label: qd.label.as_str().into(),
+            self_time: qd.self_time.as_nanos() as u64,
+            percent_total_time: ((qd.self_time.as_secs_f64() / total_time.as_secs_f64()) * 100.0)
+                as f32,
+            number_of_cache_misses: qd.number_of_cache_misses as u32,
+            number_of_cache_hits: qd.number_of_cache_hits as u32,
+            invocation_count: qd.invocation_count as u32,
+            blocked_time: qd.blocked_time.as_nanos() as u64,
+            incremental_load_time: qd.incremental_load_time.as_nanos() as u64,
+        })
+        .collect();
+
+    let totals = self_profile::QueryData {
+        label: "Totals".into(),
+        self_time: total_time.as_nanos() as u64,
+        // TODO: check against wall-time from perf stats
+        percent_total_time: cpu_clock
+            .map(|w| ((total_time.as_secs_f64() / w) * 100.0) as f32)
+            // sentinel "we couldn't compute this time"
+            .unwrap_or(-100.0),
+        number_of_cache_misses: profile
+            .query_data
+            .iter()
+            .map(|qd| qd.number_of_cache_misses as u32)
+            .sum(),
+        number_of_cache_hits: profile
+            .query_data
+            .iter()
+            .map(|qd| qd.number_of_cache_hits as u32)
+            .sum(),
+        invocation_count: profile
+            .query_data
+            .iter()
+            .map(|qd| qd.invocation_count as u32)
+            .sum(),
+        blocked_time: profile
+            .query_data
+            .iter()
+            .map(|qd| qd.blocked_time.as_nanos() as u64)
+            .sum(),
+        incremental_load_time: profile
+            .query_data
+            .iter()
+            .map(|qd| qd.incremental_load_time.as_nanos() as u64)
+            .sum(),
+    };
+
+    let artifact_sizes = profile
+        .artifact_sizes
+        .iter()
+        .map(|a| ArtifactSize {
+            label: a.label.as_str().into(),
+            bytes: a.value,
+        })
+        .collect();
+
+    Ok(self_profile::SelfProfile {
+        query_data,
+        totals,
+        artifact_sizes: Some(artifact_sizes),
+    })
 }
 
 fn extract(compressed: &[u8]) -> anyhow::Result<Vec<u8>> {
